Fix bugs with plNetAddress.
This allows the patcher to successfully connect to both GateKeeper and
FileSrv.

+++ Sources/Plasma/NucleusLib/pnUtils/pnUtAddr.h
@@ -79,7 +79,7 @@ typedef uint32_t NetAddressNode;
 bool NetAddressFromString (
     NetAddress *    addr,
     const wchar_t     str[],
-    unsigned        defaultPort
+    uint16_t        defaultPort
 );
 
 #endif
